chore: application multi-modularization
