--- src/test/kotlin/com/baek9/messaing/kafka/MessageFormDeserializerTest.kt
@@ -1,6 +1,0 @@
-package com.baek9.messaing.kafka
-
-import org.junit.jupiter.api.Assertions.*
-
-class MessageFormDeserializerTest
-
